Update README.md
The "Prime Time App" is a Scala-based interactive program that offers an engaging and educational experience. It generates a curated list of prime numbers between 201 and 499 and challenges users to identify prime numbers among randomly generated figures within the same range. The user-friendly interface facilitates simple interactions. Users can input "Y" or "N" to indicate whether a number is prime or not. The program provides immediate feedback on the correctness of the answers. This project is an excellent way to reinforce understanding of prime numbers while enjoying an interactive and entertaining game, making it a valuable educational tool for users of all ages.